pyproject.toml
[build-system]
requires = ["setuptools>=65.5.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "conversational-cv-builder"
version = "0.1.0"
description = "Conversational CV building platform"
authors = [
  { name = "NTT DATA", email = "venkatakirankumar.janga@nttdata.com" }
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.110.0",
  "uvicorn[standard]>=0.24.0",
  "pydantic>=2.6.0",
  "pydantic-settings>=2.0.0",
  "loguru>=0.7.0",
  "python-dotenv>=1.0.0",
  "httpx>=0.26.0",
  "pytest>=8.0.0",
  "pytest-asyncio>=0.23.0",
  "python-docx>=0.8.11",
  "PyPDF2>=3.0.0",
  "python-multipart>=0.0.6",
  "reportlab>=4.0.0"
]

[tool.pytest]
asyncio_mode = "auto"
